Ceramic/fibre recipes and worktables, Descriptions, Graphics
- MainTabs descriptions changed (Research/Architect)
- Recipes for AdobeBricks, ShapeAmphora, WeaveCottonCloth, WeaveLinen,
FireAmphora, FireMudBricks
- ExamineFlora now requires BastFibre
- StuffCategory Bricky
- ThingCategory Bricks (mud bricks, fired mud bricks)
- Stubs TableLoom, TablePottery (no graphics)
- Hay description
- RawCotton became CottonLint + description, graphic
- LinenPulp from Flax, Linen from Flax
- Mud (dryable to bricks), Clay clumps + graphics
- BricksAdobe (dryable) = Mud + Hay
- Mudbricks from drying Mud/BricksAdobe
- Fired mud bricks from firing Mud/BricksAdobe/Mudbricks
- Neolithic broom has a color
- UnfiredAmphora (firable in Kiln), untradeable
- HayGrass drops 18x Grass
- PlantCotton drops 5x CottonLint (25 to make Cloth)
- PlantFlax (new) drops 5x Flax (dryable to pulp, 25 to make Linen)
- Moved GameStartDialog (RA_GameStartDialog) from C# source code to a
Language file Misc.xml, C# now refers to language file
- Misc_Gameplay.xml, added DryingDescription (fixed)
- CompDryable now multiplies float * float for days of drying left
- Bricks.png, Clumps.png, Clumps_Adobe.png
- Cotton.png, Flax.png (raw resources)
- UnfinishedCloth.png (linen/cotton unfinished jobs, they take 12000
work)
- UnfinishedPottery.png (amphora unfinished job, 12000 work)
- FlaxPlant.png
- ThingCategories\Bricks.png

+++ Source/Core/Things/Components/CompDryable.cs
@@ -167,7 +167,7 @@ namespace RA
 		{
 			get
 			{
-				return Mathf.RoundToInt(daysToDry * GenDate.TicksPerDay);
+				return Mathf.RoundToInt(daysToDry * (float)GenDate.TicksPerDay);
 			}
 		}
